+++ tutorial6/main.py
@@ -1,18 +1,7 @@
-import strawberry
+﻿import strawberry
 from fastapi import FastAPI
 from strawberry.fastapi import GraphQLRouter
-
-
-@strawberry.type
-class Query:
-    @strawberry.field(description="hello string")
-    def hello(self) -> str:
-        return "hello world!!!"
-
-    @strawberry.field(description="bye string")
-    def bye(self) -> str:
-        return "good bye!!"
-
+from .mapping import Query
 
 schema = strawberry.Schema(query=Query)
 graphql_app = GraphQLRouter(schema)
