+++ tutorial9/mapping.py
@@ -1,12 +1,12 @@
 import strawberry
-from schema import FootballPlayer, CricketPlayer, Invalid
-from data import read_data
+from .schema import FootballPlayer, CricketPlayer, Invalid, Player
+from .data import read_data
 
 
 @strawberry.type
 class Query:
     @strawberry.field
-    def player(self, playertype: str) -> list:
+    def player(self, playertype: str) -> list[Player]:
         data = read_data()
         if playertype == "fplayer":
             return [FootballPlayer(**p) for p in data[0]["footballplayer"]]
