+++ tutorial9/schema.py
@@ -1,25 +1,23 @@
 import strawberry
 
 
-@strawberry.type
-class FootballPlayer:
+@strawberry.interface
+class Player:
     name: str
     country: str
     type: str
+
+
+@strawberry.type
+class FootballPlayer(Player):
     position: str
 
 
 @strawberry.type
-class CricketPlayer:
-    name: str
-    country: str
-    type: str
+class CricketPlayer(Player):
     battingorder: int
 
 
 @strawberry.type
-class Invalid:
-    name: str
-    country: str
-    type: str
+class Invalid(Player):
     invalid_data: str
